code refactory  dir
